Updated user and task model and tagging UI
Added due_date and due_time fields for task model and UI. Updated tagging UI to
allow for deletion of tags. Changed the update button in show/edit
fields to write to db immediately. Updated user model so that it
recognises when a user is using the app as a guest. Changed the
controller for login such that tasks as a guest is written to the user's
account on account creation.